Report Changes + Spring Security (Logout Changes)

+++ GramYojana/src/com/daminfo/dao/BlockDAO.java
@@ -2,12 +2,12 @@ package com.daminfo.dao;
 
 import java.util.List;
 
+import com.daminfo.common.json.Location;
 import com.daminfo.model.Block;
-import com.daminfo.service.common.Location;
 
 public interface BlockDAO {
 
-	List<Location> getAllBlocksByDistrictID(int districtCode);
-	Block getBlockInfo(int blockCode);
+	List<Location> getAllBlocksByDistrictID(long districtCode);
+	Block getBlockInfo(long blockCode);
 
 }
