+++ GramYojana/src/com/daminfo/service/impl/LocationServiceImpl.java
@@ -3,13 +3,13 @@ package com.daminfo.service.impl;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.daminfo.common.json.LocationList;
 import com.daminfo.dao.BlockDAO;
 import com.daminfo.dao.DistrictDAO;
 import com.daminfo.dao.PanchayatDAO;
 import com.daminfo.dao.StateDAO;
 import com.daminfo.dao.VillageDAO;
-import com.daminfo.service.common.LocationList;
-import com.daminfo.service.common.LocationService;
+import com.daminfo.service.LocationService;
 
 @Service
 public class LocationServiceImpl implements LocationService{
@@ -50,7 +50,7 @@ public class LocationServiceImpl implements LocationService{
 
 	@Override
 	@Transactional
-	public LocationList getDistrictList(int stateCode) {
+	public LocationList getDistrictList(long stateCode) {
 		LocationList locList = new LocationList();
 		locList.setLocations(this.districtDAO.getAllDistrictsByStateID(stateCode));
 		return locList;
@@ -58,7 +58,7 @@ public class LocationServiceImpl implements LocationService{
 
 	@Override
 	@Transactional
-	public LocationList getBlockList(int districtCode) {
+	public LocationList getBlockList(long districtCode) {
 		LocationList locList = new LocationList();
 		locList.setLocations(this.blockDAO.getAllBlocksByDistrictID(districtCode));
 		return locList;
@@ -66,7 +66,7 @@ public class LocationServiceImpl implements LocationService{
 
 	@Override
 	@Transactional
-	public LocationList getPanchayatList(int blockCode) {
+	public LocationList getPanchayatList(long blockCode) {
 		LocationList locList = new LocationList();
 		locList.setLocations(this.panchayatDAO.getAllPanchayatsByBlockID(blockCode));
 		return locList;
@@ -74,7 +74,7 @@ public class LocationServiceImpl implements LocationService{
 
 	@Override
 	@Transactional
-	public LocationList getVillageList(int panchayatCode) {
+	public LocationList getVillageList(long panchayatCode) {
 		LocationList locList = new LocationList();
 		locList.setLocations(this.villageDAO.getAllVillagesByPanchayatID(panchayatCode));
 		return locList;
